app/requirements.txt
uvicorn==0.13.3
pytesseract==0.3.7
fastapi==0.63.0
aiofiles==0.6.0
Wand==0.6.6
Pillow==8.1.2
jinja2==2.11.3
python-multipart==0.0.5
# fastapi==0.65.2